requirements.txt
fastapi
uvicorn
face_recognition
opencv-python
opencv-contrib-python
pillow
numpy
mysql-connector-python
